Rewrite configuration in Dhall.
This helps keep the configuration DRY-er and will save us from some
repetition of file paths when we add Stmt classes to the AST later.

Of course, we could just give Deno write access to the entire "bourbon"
directory, but why have a fine-grained permission system if we're not
going to use it?